+++ lib/tty/progressbar/configuration.rb
@@ -11,11 +11,11 @@ module TTY
 
       attr_accessor :width
 
-      attr_accessor :incomplete
+      attr_reader :incomplete
 
-      attr_accessor :complete
+      attr_reader :complete
 
-      attr_accessor :unknown
+      attr_reader :unknown
 
       attr_accessor :head
 
@@ -39,9 +39,9 @@ module TTY
         self.total   = options[:total] if options[:total]
         @width       = options.fetch(:width) { total }
         @bar_format  = options.fetch(:bar_format, :classic)
-        @incomplete  = options.fetch(:incomplete) { fetch_char(@bar_format, :incomplete) }
-        @complete    = options.fetch(:complete) { fetch_char(@bar_format, :complete) }
-        @unknown     = options.fetch(:unknown) { fetch_char(@bar_format, :unknown) }
+        self.incomplete = options.fetch(:incomplete) { fetch_char(@bar_format, :incomplete) }
+        self.complete = options.fetch(:complete) { fetch_char(@bar_format, :complete) }
+        self.unknown = options.fetch(:unknown) { fetch_char(@bar_format, :unknown) }
         @head        = options.fetch(:head) { @complete || "=" }
         @clear_head  = options.fetch(:clear_head, false)
         @hide_cursor = options.fetch(:hide_cursor, false)
@@ -50,6 +50,39 @@ module TTY
         @frequency   = options.fetch(:frequency, 0) # 0Hz
         @interval    = options.fetch(:interval, 1) # 1 sec
         @inset       = options.fetch(:inset, 0)
+      end
+
+      # Set complete character(s)
+      #
+      # @param [String] value
+      #
+      # @api public
+      def complete=(value)
+        raise_if_empty(:complete, value)
+
+        @complete = value
+      end
+
+      # Set incomplete character(s)
+      #
+      # @param [String] value
+      #
+      # @api public
+      def incomplete=(value)
+        raise_if_empty(:incomplete, value)
+
+        @incomplete = value
+      end
+
+      # Set unknown character(s)
+      #
+      # @param [String] value
+      #
+      # @api public
+      def unknown=(value)
+        raise_if_empty(:unknown, value)
+
+        @unknown = value
       end
 
       def total=(value)
@@ -75,6 +108,17 @@ module TTY
                                "Available formats are: " \
                                "#{FORMATS.keys.sort.map(&:inspect).join(', ')}"
         end
+      end
+
+      # Check whether a parameter's value is empty or not
+      #
+      # @raise [ArgumentError]
+      #
+      # @api private
+      def raise_if_empty(name, value)
+        return value unless value.to_s.empty?
+
+        raise ArgumentError, "cannot provide an empty string for #{name.inspect}"
       end
     end # Configuration
   end # ProgressBar