+++ spec/unit/configure_spec.rb
@@ -21,4 +21,28 @@ RSpec.describe TTY::ProgressBar, "#configure" do
 
     expect(progress.total).to eq(20)
   end
+
+  %i[complete incomplete unknown].each do |option|
+    it "fails to set #{option.inspect} to an empty string via initialize" do
+      expect {
+        described_class.new(":bar", total: 10, option => "")
+      }.to raise_error(
+        ArgumentError,
+        "cannot provide an empty string for #{option.inspect}"
+      )
+    end
+
+    it "fails to set #{option.inspect} to empty string via accessor" do
+      progress = described_class.new(":bar", total: 10)
+
+      expect {
+        progress.configure do |config|
+          config.public_send("#{option}=", "")
+        end
+      }.to raise_error(
+        ArgumentError,
+        "cannot provide an empty string for #{option.inspect}"
+      )
+    end
+  end
 end