+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 * Change to update strings-ansi and tty-screen dependencies
 * Change Pipeline to inject progress bar instance only once
 * Change :elapsed and :eta to show days after running for 24 hours
+* Change to ensure complete, incomplete and unknown option cannot be an empty string
 
 ### Fixed
 * Fix MultiBar top bar to allow resuming progress when stopped/done (@d4be4st)